Изменение типа id и обобщение сущностей при помощи Essence в entities

+++ excel_processing/src/main/java/com/shirayev/excel_processing/repositories/PeoplePassageRepository.java
@@ -3,5 +3,5 @@ package com.shirayev.excel_processing.repositories;
 import com.shirayev.excel_processing.entities.PeoplePassage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PeoplePassageRepository extends JpaRepository<PeoplePassage, Integer> {
+public interface PeoplePassageRepository extends JpaRepository<PeoplePassage, Long> {
 }
